requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.3.0
fastapi==0.117.1
h11==0.16.0
idna==3.10
numpy==2.3.2
pandas==2.3.1
py4j==0.10.9.9
pydantic==2.11.9
pydantic_core==2.33.2
pyodbc==5.2.0
pyspark==4.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.36.0
